inference/pyproject.toml
[project]
name = "image-captioning-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pillow>=12.0.0",
]
